+++ NEWS.md
@@ -1,8 +1,9 @@
-# nanonext 0.8.1.9003 (development)
+# nanonext 0.8.1.9004 (development)
 
 #### New Features
 
-* `'cv_value<-'()` implemented to set the value of a condition value, providing greater flexibility.
+* `'cv_value<-'()` implemented to set the value of a condition value, providing greater flexibility in use.
+* `lock()` and `unlock()` implemented to prevent further pipe connections from being established at a socket, optionally tied to the value of a condition variable.
 
 #### Updates